Attempting to fix vine
