+++ Botanist-Journey/Assets/Prefab/Vine.prefab
@@ -14,11 +14,10 @@ GameObject:
   - component: {fileID: 1329915041367785809}
   - component: {fileID: 4520560860716128624}
   - component: {fileID: 6102784836947723831}
-  - component: {fileID: 8545875231293576388}
-  - component: {fileID: 8029243056993083553}
+  - component: {fileID: 5348603251491160687}
   m_Layer: 0
   m_Name: Vine
-  m_TagString: Untagged
+  m_TagString: Plant
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -250,7 +249,7 @@ MonoBehaviour:
   m_PreserveMeshAssetOnDestroy: 0
   assetGuid: 
   m_Mesh: {fileID: 0}
-  m_VersionIndex: 1428
+  m_VersionIndex: 1440
   m_IsSelectable: 1
   m_SelectedFaces: 
   m_SelectedEdges: []
@@ -351,7 +350,7 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1.2308663, y: 4, z: 0.46481323}
   m_Center: {x: 0.015677392, y: 1.5, z: 0.26759338}
---- !u!114 &8545875231293576388
+--- !u!114 &5348603251491160687
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -360,27 +359,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 1098999416354492594}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ab7da97958dc9c541b3bcf9e1c63bc5a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  climbSpeed: 2
-  player: {fileID: 0}
---- !u!114 &8029243056993083553
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1098999416354492594}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f5e8628e6e50c34428350ff06496a0e2, type: 3}
+  m_Script: {fileID: 11500000, guid: 314fee0c22ef7934b9d27fa4ecadcce9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 0}
   requiredTool:
-    itemName: 
+    itemName: machete
     itemDescription: 
     plantType: 0
     toolType: 1
-  itemPrefab: {fileID: 0}
+  itemPickupPrefab: {fileID: 1098999416354492594}
