+++ Botanist-Journey/Assets/Scripts/Interaction.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 314fee0c22ef7934b9d27fa4ecadcce9
