+++ Botanist-Journey/Assets/Scripts/ItemHolder.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class ItemHolder: MonoBehaviour
+{
+    public Item item;
+}
