--- Botanist-Journey/Assets/Scripts/PlantInteraction.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f5e8628e6e50c34428350ff06496a0e2
