+++ Botanist-Journey/ProjectSettings/TagManager.asset
@@ -5,6 +5,8 @@ TagManager:
   serializedVersion: 3
   tags:
   - Ground
+  - Plant
+  - Tool
   layers:
   - Default
   - TransparentFX
